directions 타입 정리, /about/directions SSR로 변경
