--- types/directions.ts
@@ -1,5 +1,0 @@
-export interface Direction {
-  name: string;
-  engName: string;
-  description: string;
-}
